feat(nextbike/prediction): renew structure of files
