+++ nextbike/prediction/__init__.py
@@ -1,1 +1,3 @@
 from .training import *
+from .split import *
+from .predict import *
